Rename package from com.github.fkrull.konflin to just konflin
