--- common/src/main/kotlin/com/github/fkrull/konflin/ConfigurationSource.kt
@@ -1,7 +1,0 @@
-package com.github.fkrull.konflin
-
-interface ConfigurationSource {
-    fun getString(name: String): String?
-    fun getInt(name: String): Int?
-    fun getBoolean(name: String): Boolean?
-}
