--- common/src/main/kotlin/com/github/fkrull/konflin/MissingConfigValueException.kt
@@ -1,3 +1,0 @@
-package com.github.fkrull.konflin
-
-class MissingConfigValueException(message: String) : RuntimeException(message)
