--- common/src/main/kotlin/com/github/fkrull/konflin/Setting.kt
@@ -1,15 +1,0 @@
-package com.github.fkrull.konflin
-
-interface Setting<out T> {
-    val name: String
-
-    /**
-     * @suppress
-     */
-    val default: T?
-
-    /**
-     * @suppress
-     */
-    fun get(configSource: ConfigurationSource): T?
-}
