--- common/src/main/kotlin/com/github/fkrull/konflin/UnsupportedConfigTypeException.kt
@@ -1,3 +1,0 @@
-package com.github.fkrull.konflin
-
-class UnsupportedConfigTypeException(message: String) : IllegalArgumentException(message)
