--- common/src/main/kotlin/com/github/fkrull/konflin/converter/ConfigType.kt
@@ -1,11 +1,0 @@
-package com.github.fkrull.konflin.converter
-
-import kotlin.reflect.KClass
-
-sealed class ConfigType<out Type : Any>(val clazz: KClass<out Type>) {
-    companion object Types {
-        object String : ConfigType<kotlin.String>(kotlin.String::class)
-        object Int : ConfigType<kotlin.Int>(kotlin.Int::class)
-        object Boolean : ConfigType<kotlin.Boolean>(kotlin.Boolean::class)
-    }
-}
