--- common/src/main/kotlin/com/github/fkrull/konflin/converter/impl/IdentityConverter.kt
@@ -1,9 +1,0 @@
-package com.github.fkrull.konflin.converter.impl
-
-import com.github.fkrull.konflin.converter.ConfigType
-import com.github.fkrull.konflin.converter.Converter
-
-internal class IdentityConverter<Type : Any>(override val inType: ConfigType<Type>) : Converter<Type, Type> {
-    override val outType = inType.clazz
-    override fun fromConfig(value: Type): Type = value
-}
