--- common/src/main/kotlin/com/github/fkrull/konflin/impl/KClassId.kt
@@ -1,7 +1,0 @@
-package com.github.fkrull.konflin.impl
-
-import kotlin.reflect.KClass
-
-internal expect class KClassId(value: Any)
-
-internal expect val <T : Any> KClass<T>.id: KClassId
