+++ common/src/main/kotlin/konflin/ConfigurationSource.kt
@@ -1,0 +1,7 @@
+package konflin
+
+interface ConfigurationSource {
+    fun getString(name: String): String?
+    fun getInt(name: String): Int?
+    fun getBoolean(name: String): Boolean?
+}
