+++ common/src/main/kotlin/konflin/MissingConfigValueException.kt
@@ -1,0 +1,3 @@
+package konflin
+
+class MissingConfigValueException(message: String) : RuntimeException(message)
