+++ common/src/main/kotlin/konflin/Setting.kt
@@ -1,0 +1,15 @@
+package konflin
+
+interface Setting<out T> {
+    val name: String
+
+    /**
+     * @suppress
+     */
+    val default: T?
+
+    /**
+     * @suppress
+     */
+    fun get(configSource: ConfigurationSource): T?
+}
