+++ common/src/main/kotlin/konflin/UnsupportedConfigTypeException.kt
@@ -1,0 +1,3 @@
+package konflin
+
+class UnsupportedConfigTypeException(message: String) : IllegalArgumentException(message)
