+++ common/src/main/kotlin/konflin/converter/Converter.kt
@@ -1,0 +1,12 @@
+package konflin.converter
+
+import kotlin.reflect.KClass
+
+// Type parameter is not `out` for future extension of a `toConfig` method.
+@Suppress("AddVarianceModifier")
+interface Converter<OutType : Any, InType : Any> {
+    val inType: ConfigType<InType>
+    val outType: KClass<out OutType>
+
+    fun fromConfig(value: InType): OutType
+}
