+++ common/src/main/kotlin/konflin/impl/KClassId.kt
@@ -1,0 +1,7 @@
+package konflin.impl
+
+import kotlin.reflect.KClass
+
+internal expect class KClassId(value: Any)
+
+internal expect val <T : Any> KClass<T>.id: KClassId
