+++ jvm/src/main/kotlin/konflin/impl/KClassId.kt
@@ -1,0 +1,7 @@
+package konflin.impl
+
+import kotlin.reflect.KClass
+
+internal actual data class KClassId actual constructor(val value: Any)
+
+internal actual val <T : Any> KClass<T>.id: KClassId get() = KClassId(this)
